Make moving around smoother
Go only responds with a world packet when the player logs in. In the
future, send this when we need to load chunks. Player info is sent
on every player move. On the CS side, the world is only redrawn when
a world packet is received, otherwise we just move the graphics object
around.